refactor(NamespaceSelectionAgent): extract utility functions into separate modules
Extract helper methods from NamespaceSelectionAgent into single-purpose modules:
- namespace_data.py: models, context keys, data fetching, RAG helper
- namespace_formatting.py: prompt formatting utilities
- selection_parsing.py: parsing models and LLM structured output

Reduces agent class from 443 to 219 lines, improving readability and testability.
